Customer Churn Prediction/deployment/requirements.txt
streamlit
pandas
tensorflow-cpu
sklearn
